backend/requirements.txt
# ============================================================
# RenAIssance OCR Backend — Python dependencies
# Pinned for reproducibility with Python 3.11
# ============================================================

# --- Web framework ---
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.4

# --- HTTP client (ChatGPT, DeepSeek, Qwen providers) ---
httpx==0.28.1

# --- AI / OCR providers ---
google-genai==1.0.0
protobuf==5.29.2

# --- Document export ---
python-docx==1.1.2
reportlab==4.2.5

# --- Image processing ---
Pillow==11.1.0
opencv-python-headless==4.10.0.84
numpy==1.26.4

# --- Utilities ---
python-dotenv==1.0.1

# --- PDF parsing (export service) ---
PyMuPDF==1.25.1

# --- Layout-aware detection (optional heavy deps) ---
# Comment these out if not using layout detection feature
# to significantly reduce image size (~3 GB saved)
paddlepaddle==3.0.0
paddleocr==2.10.0